Implement new flow to play audio-only media assets
This is not replacing the previous one, but adding an extra way that can be changed in the AndroidManifest.xml file

With the new implementation, now it is possible to load a list of media and displayed on the screen. By tapping one of the media items, it will load a second screen with just the player to load and play the media.